requirements.txt
# local package
-e .

# external requirements
click
coverage
flake8
python-dateutil
python-dotenv>=0.5.1
requests
jupyter